LayoutTests/webgpu/whlsl/int-literal-compare.html

<!DOCTYPE html>
<html>
<meta charset=utf-8>
<meta name="timeout" content="long">
<title>Test the WHLSL test harness.</title>
<script src="js/test-harness.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
const whlslTests = {};

whlslTests.twoIntLiterals = async () => {
    let program = `
        bool foo()
        {
            return 42 == 42;
        }
    `;
    assert_equals(await callBoolFunction(program, "foo", []), true);
}

runTests(whlslTests);
</script>
</html>
